services/feature-svc/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.0
torch==2.2.0
git+https://github.com/openai/CLIP.git
Pillow==10.2.0
numpy==1.26.4
python-multipart==0.0.6
boto3==1.34.25
pandas==2.2.0
pyarrow==15.0.0
